DS-2: Modify structure and java docs for Collection unit tests. #3
