+++ data-structures/src/test/java/com/github/vmalaya/sigmasoftware/internship/datastructures/iterable/collection/CollectionIsEmptyTest.java
@@ -1,0 +1,65 @@
+package com.github.vmalaya.sigmasoftware.internship.datastructures.iterable.collection;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.*;
+import java.util.concurrent.*;
+
+/**
+ * All collections can determine that they don't contain any element:
+ * List:
+ *   ArrayList
+ *   LinkedList
+ *   Vector
+ *     Stack
+ * Set:
+ *   HashSet
+ *   LinkedHashSet
+ *   TreeSet
+ *   ConcurrentSkipListSet
+ *   RegularEnumSet
+ *   CopyOnWriteArraySet
+ * Queue:
+ *   ConcurrentLinkedQueue
+ *   PriorityQueue
+ *   LinkedBlockingDeque
+ *   ArrayDeque
+ *   ConcurrentLinkedDeque
+ *   ArrayBlockingQueue
+ *   LinkedTransferQueue
+ *   LinkedBlockingQueue
+ *   PriorityBlockingQueue
+ *   LinkedTransferQueue
+ */
+public class CollectionIsEmptyTest {
+
+    @Test
+    public void should_check_for_emptiness() {
+
+        //given
+        int capacity = 1;
+        List<? extends Collection<Object>> collections = Arrays.asList(new ArrayList<>(),
+                new LinkedList<>(),
+                new Vector<>(),
+                new Stack<>(),
+                new CopyOnWriteArrayList<>(),
+                new ArrayBlockingQueue<>(capacity),
+                new PriorityQueue<>(),
+                new ConcurrentLinkedQueue<>(),
+                new LinkedBlockingQueue<>(),
+                new LinkedTransferQueue<>(),
+                new ArrayDeque<>(),
+                new ConcurrentLinkedDeque<>(),
+                new LinkedBlockingDeque<>(),
+                new HashSet<>(),
+                new LinkedHashSet<>(),
+                new ConcurrentSkipListSet<>(),
+                new PriorityBlockingQueue<>(),
+                new TreeSet<>(),
+                new CopyOnWriteArraySet<>());
+
+        //then
+        collections.forEach(collection -> Assert.assertTrue(collection.isEmpty()));
+    }
+}
