+++ data-structures/src/test/java/com/github/vmalaya/sigmasoftware/internship/datastructures/iterable/collection/CollectionRetainTest.java
@@ -1,0 +1,71 @@
+package com.github.vmalaya.sigmasoftware.internship.datastructures.iterable.collection;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.*;
+import java.util.concurrent.*;
+
+/**
+ * All collections can retain only elements of given collection:
+ * List:
+ *   ArrayList
+ *   LinkedList
+ *   Vector
+ *     Stack
+ * Set:
+ *   HashSet
+ *   LinkedHashSet
+ *   TreeSet
+ *   ConcurrentSkipListSet
+ *   RegularEnumSet
+ *   CopyOnWriteArraySet
+ * Queue:
+ *   ConcurrentLinkedQueue
+ *   PriorityQueue
+ *   LinkedBlockingDeque
+ *   ArrayDeque
+ *   ConcurrentLinkedDeque
+ *   ArrayBlockingQueue
+ *   LinkedTransferQueue
+ *   LinkedBlockingQueue
+ *   PriorityBlockingQueue
+ *   LinkedTransferQueue
+ */
+public class CollectionRetainTest {
+
+    @Test
+    public void should_retain_only_contained_elements_in_collection() {
+
+        //given
+        int capacity = 2;
+        List<? extends Collection<Object>> collections = Arrays.asList(new ArrayList<>(),
+                new LinkedList<>(),
+                new Vector<>(),
+                new Stack<>(),
+                new CopyOnWriteArrayList<>(),
+                new ArrayBlockingQueue<>(capacity),
+                new PriorityQueue<>(),
+                new ConcurrentLinkedQueue<>(),
+                new LinkedBlockingQueue<>(),
+                new LinkedTransferQueue<>(),
+                new ArrayDeque<>(),
+                new ConcurrentLinkedDeque<>(),
+                new LinkedBlockingDeque<>(),
+                new HashSet<>(),
+                new LinkedHashSet<>(),
+                new ConcurrentSkipListSet<>(),
+                new PriorityBlockingQueue<>(),
+                new TreeSet<>(),
+                new CopyOnWriteArraySet<>());
+
+        //and
+        collections.forEach(collection -> collection.addAll(Arrays.asList("string1", "string2")));
+
+        //when
+        collections.forEach(collection -> collection.retainAll(Arrays.asList("string1")));
+
+        //then
+        collections.forEach(collection -> Assert.assertFalse(collection.contains("string2")));
+    }
+}
