+++ data-structures/src/test/java/com/github/vmalaya/sigmasoftware/internship/datastructures/iterable/collection/CollectionStreamTest.java
@@ -1,13 +1,14 @@
-package com.github.vmalaya.sigmasoftware.internship.datastructures.iterable;
+package com.github.vmalaya.sigmasoftware.internship.datastructures.iterable.collection;
 
 import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.*;
 import java.util.concurrent.*;
+import java.util.stream.Stream;
 
 /**
- * All collections are iterable:
+ * All collections can return stream:
  * List:
  *   ArrayList
  *   LinkedList
@@ -32,39 +33,10 @@ import java.util.concurrent.*;
  *   PriorityBlockingQueue
  *   LinkedTransferQueue
  */
-public class IterableTest {
+public class CollectionStreamTest {
 
     @Test
-    public void should_do_action_for_each_element() {
-
-        //given
-        int capacity = 1;
-        List<? extends Collection<Object>> collections = Arrays.asList(new ArrayList<>(),
-                new LinkedList<>(),
-                new Vector<>(),
-                new Stack<>(),
-                new CopyOnWriteArrayList<>(),
-                new ArrayBlockingQueue<>(capacity),
-                new PriorityQueue<>(),
-                new ConcurrentLinkedQueue<>(),
-                new LinkedBlockingQueue<>(),
-                new LinkedTransferQueue<>(),
-                new ArrayDeque<>(),
-                new ConcurrentLinkedDeque<>(),
-                new LinkedBlockingDeque<>(),
-                new HashSet<>(),
-                new LinkedHashSet<>(),
-                new ConcurrentSkipListSet<>(),
-                new PriorityBlockingQueue<>(),
-                new TreeSet<>(),
-                new CopyOnWriteArraySet<>());
-
-        //then
-        collections.forEach(collection -> Assert.assertTrue(true));
-    }
-
-    @Test
-    public void should_return_iterator() {
+    public void should_return_stream() {
 
         //given
         int capacity = 1;
@@ -90,11 +62,11 @@ public class IterableTest {
 
         //then
         collections.forEach(collection ->
-                Assert.assertTrue(collection.iterator() instanceof Iterator));
+                Assert.assertTrue(collection.stream() instanceof Stream));
     }
 
     @Test
-    public void should_return_spliterator() {
+    public void should_return_parallel_stream() {
 
         //given
         int capacity = 1;
@@ -120,6 +92,6 @@ public class IterableTest {
 
         //then
         collections.forEach(collection ->
-                Assert.assertTrue(collection.spliterator() instanceof Spliterator));
+                Assert.assertTrue(collection.parallelStream().isParallel()));
     }
 }
