--- data-structures/src/test/java/com/github/vmalaya/sigmasoftware/internship/datastructures/iterable/collection/ConvertingToArray.java
@@ -1,155 +1,0 @@
-package com.github.vmalaya.sigmasoftware.internship.datastructures.iterable.collection;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.util.*;
-import java.util.concurrent.*;
-
-/**
- * Testing following classes as collection for converting to Array:
- * ArrayList
- * LinkedList
- * Vector
- * Stack
- * HashSet
- * <p>
- * LinkedHashSet
- * TreeSet
- * ConcurrentSkipListSet
- * RegularEnumSet
- * CopyOnWriteArraySet
- * <p>
- * ConcurrentLinkedQueue
- * PriorityQueue
- * LinkedBlockingDeque
- * ArrayDeque
- * ConcurrentLinkedDeque
- * ArrayBlockingQueue
- * LinkedTransferQueue
- * LinkedBlockingQueue
- * DelayQueue
- * PriorityBlockingQueue
- * SynchronousQueue
- * LinkedTransferQueue
- */
-public class ConvertingToArray {
-
-    @Test
-    public void should_convert_to_array_of_objects() {
-
-        //given
-        int capacity = 1;
-        ArrayList<Object> arrayList = new ArrayList<>();
-        LinkedList<Object> linkedList = new LinkedList<>();
-        Vector<Object> vector = new Vector<>();
-        Stack<Object> stack = new Stack<>();
-        CopyOnWriteArrayList<Object> copyOnWriteArrayList = new CopyOnWriteArrayList<>();
-        ArrayBlockingQueue<Object> arrayBlockingQueue = new ArrayBlockingQueue<>(capacity);
-        PriorityQueue<Object> priorityQueue = new PriorityQueue<>();
-        ConcurrentLinkedQueue<Object> concurrentLinkedQueue = new ConcurrentLinkedQueue<>();
-        LinkedBlockingQueue<Object> linkedBlockingQueue = new LinkedBlockingQueue<>();
-        LinkedTransferQueue<Object> linkedTransferQueue = new LinkedTransferQueue<>();
-        ArrayDeque<Object> arrayDeque = new ArrayDeque<>();
-        ConcurrentLinkedDeque<Object> concurrentLinkedDeque = new ConcurrentLinkedDeque<>();
-        LinkedBlockingDeque<Object> linkedBlockingDeque = new LinkedBlockingDeque<>();
-        HashSet<Object> hashSet = new HashSet<>();
-        LinkedHashSet<Object> linkedHashSet = new LinkedHashSet<>();
-        ConcurrentSkipListSet<Object> concurrentSkipListSet = new ConcurrentSkipListSet<>();
-        PriorityBlockingQueue priorityBlockingQueue = new PriorityBlockingQueue<>();
-        TreeSet treeSet = new TreeSet<>();
-        CopyOnWriteArraySet copyOnWriteArraySet = new CopyOnWriteArraySet<>();
-        DelayQueue<Delayed> delayQueue = new DelayQueue<>();
-
-        SynchronousQueue synchronousQueue = new SynchronousQueue(); //TODO: should be tested toArray()
-
-        List<? extends Collection> collections = Arrays.asList(arrayList,
-                linkedList,
-                vector,
-                stack,
-                copyOnWriteArrayList,
-                arrayBlockingQueue,
-                priorityQueue,
-                concurrentLinkedQueue,
-                linkedBlockingQueue,
-                linkedTransferQueue,
-                arrayDeque,
-                concurrentLinkedDeque,
-                linkedBlockingDeque,
-                hashSet,
-                linkedHashSet,
-                concurrentSkipListSet,
-                priorityBlockingQueue,
-                treeSet,
-                copyOnWriteArraySet);
-        collections.forEach(n -> n.add(true));
-        delayQueue.add(new DelayObject("TEST", 1));
-
-        //when
-        Object[] objects = collections.toArray();
-        Object[] delayObject = delayQueue.toArray();
-
-        //then
-        Assert.assertTrue(objects instanceof Object);
-        Assert.assertTrue(delayObject instanceof Object);
-    }
-
-    @Test
-    public void should_convert_to_array_of_given_type() {
-
-        //given
-        int capacity = 1;
-        ArrayList<Object> arrayList = new ArrayList<>();
-        LinkedList<Object> linkedList = new LinkedList<>();
-        Vector<Object> vector = new Vector<>();
-        Stack<Object> stack = new Stack<>();
-        CopyOnWriteArrayList<Object> copyOnWriteArrayList = new CopyOnWriteArrayList<>();
-        ArrayBlockingQueue<Object> arrayBlockingQueue = new ArrayBlockingQueue<>(capacity);
-        PriorityQueue<Object> priorityQueue = new PriorityQueue<>();
-        ConcurrentLinkedQueue<Object> concurrentLinkedQueue = new ConcurrentLinkedQueue<>();
-        LinkedBlockingQueue<Object> linkedBlockingQueue = new LinkedBlockingQueue<>();
-        LinkedTransferQueue<Object> linkedTransferQueue = new LinkedTransferQueue<>();
-        ArrayDeque<Object> arrayDeque = new ArrayDeque<>();
-        ConcurrentLinkedDeque<Object> concurrentLinkedDeque = new ConcurrentLinkedDeque<>();
-        LinkedBlockingDeque<Object> linkedBlockingDeque = new LinkedBlockingDeque<>();
-        HashSet<Object> hashSet = new HashSet<>();
-        LinkedHashSet<Object> linkedHashSet = new LinkedHashSet<>();
-        ConcurrentSkipListSet<Object> concurrentSkipListSet = new ConcurrentSkipListSet<>();
-        PriorityBlockingQueue priorityBlockingQueue = new PriorityBlockingQueue<>();
-        TreeSet treeSet = new TreeSet<>();
-        CopyOnWriteArraySet copyOnWriteArraySet = new CopyOnWriteArraySet<>();
-        DelayQueue<Delayed> delayQueue = new DelayQueue<>();
-
-        SynchronousQueue synchronousQueue = new SynchronousQueue(); //TODO: should be tested toArray()
-
-        List<? extends Collection> collections = Arrays.asList(arrayList,
-                linkedList,
-                vector,
-                stack,
-                copyOnWriteArrayList,
-                arrayBlockingQueue,
-                priorityQueue,
-                concurrentLinkedQueue,
-                linkedBlockingQueue,
-                linkedTransferQueue,
-                arrayDeque,
-                concurrentLinkedDeque,
-                linkedBlockingDeque,
-                hashSet,
-                linkedHashSet,
-                concurrentSkipListSet,
-                priorityBlockingQueue,
-                treeSet,
-                copyOnWriteArraySet);
-        collections.forEach(n -> n.add(true));
-        delayQueue.add(new DelayObject("TEST", 1));
-
-        //when
-        Collection[] objects = collections.toArray(new Collection[0]);
-        Delayed[] delayed = delayQueue.toArray(new Delayed[0]);
-
-        //then
-        Assert.assertTrue(objects instanceof Collection[]);
-        Assert.assertTrue(delayed instanceof Delayed[]);
-    }
-}
